requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.1
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.4.0
bcrypt==4.1.2
